fix destroy user service

+++ app/services/users/destroy_user_service.rb
@@ -1,5 +1,6 @@
 module Users
   class DestroyUserService
+    include ActiveModel::Validations
     attr_accessor :id, :user, :user_profile
 
     def initialize(id:)
@@ -10,39 +11,68 @@ module Users
 
     def call
       ActiveRecord::Base.transaction do
-        destroy_resources
-        user.user_profile.destroy!
-        user.destroy!
+        if destroy_resources
+          user.user_profile.destroy!
+          user.destroy!
+        end
       end
+      self
     end
 
     private
 
     def destroy_resources
       if user_profile.has_role? :student
+        destroy_student_entities(user_profile.student)
         user_profile.student.destroy!
         user_profile.remove_role :student
-      end
-
-      if user_profile.has_role? :director
-        user_profile.director.destroy!
-        user_profile.remove_role :director
+        return true
       end
 
       if user_profile.has_role? :manager
-        user_profile.manager.destroy!
-        user_profile.remove_role :manager
+        if can_manager_be_destroyed?(user_profile.manager)
+          user_profile.manager.destroy!
+          user_profile.remove_role :manager
+          return true
+        else
+          errors.add(:user, 'is a manager with active rotations')
+          return false
+        end
       end
 
       if user_profile.has_role? :professor
-        user_profile.professor.destroy!
-        user_profile.remove_role :professor
+        if can_professor_be_destroyed?(user_profile.professor)
+          user_profile.professor.destroy!
+          user_profile.remove_role :professor
+          return true
+        else
+          errors.add(:user, 'is a professor with active subjects')
+          return false
+        end
       end
 
-      if user_profile.has_role? :administrator
-        user_profile.administrator.destroy!
-        user_profile.remove_role :administrator
+      if user_profile.has_role? :superadmin
+        user_profile.remove_role :superadmin
+        return true
       end
+    end
+
+    def destroy_student_entities(student)
+      activity_califications = student.activity_califications
+      activity_califications.each { |activity_calification| activity_calification.bloom_taxonomy_levels.destroy_all if activity_calification.bloom_taxonomy_levels.present?}
+      activity_califications.destroy_all
+      student_informations = student.student_informations
+      student_informations.destroy_all
+    end
+
+    def can_manager_be_destroyed?(manager)
+      rotations = Rotation.where(manager_id: manager.id)
+      return !rotations.present?
+    end
+
+    def can_professor_be_destroyed?(professor)
+      subjects = Subject.where(professor_id: professor.id)
+      return !subjects.present?
     end
   end
 end
